requirements.txt
alembic==1.4.3
amqp==5.0.6
appdirs==1.4.3
bcrypt==3.2.0
billiard==3.6.4.0
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
html5lib==1.0.1
idna==2.8
importlib-metadata==3.10.0
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
prompt-toolkit==3.0.18
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.4.1
